refactor(auth): Plan A stages 1+5a+5b — decouple auth from main.py
Stage 1 (logger): seed_demo_people.py + seed_test_account.py no longer
import logger from main; each uses logging.getLogger(__name__) locally.

Stage 5a: routes/builder_portal_routes.py — lazy get_current_user import
now points at auth.dependencies (canonical) instead of main.

Stage 5b (core): move get_current_user + get_current_user_flexible bodies
(~310 lines) from main.py into auth/dependencies.py. Constants sourced
from auth.tokens (_USE_SECURE_TOKENS, _verify_secure_token, token_blacklist,
TokenType) and env vars — no main.py dependency at call time. Removed
_LazyAuthProxy class and _get_main_auth() entirely from auth/dependencies.py.
main.py retains a 4-line back-compat re-export so all existing
`from main import get_current_user` calls continue to work unchanged.

Verified: syntax clean on main.py + auth/dependencies.py; back-compat
re-export returns identical objects; 55 auth unit+flow tests pass, 0 new
failures (14 pre-existing integration errors are DB-env only).

Email helper scripts (sync_emails_now.py etc.) NOT READY — they depend on
init_dre_helpers() initialization that only main.py's import chain triggers.
Deferred to a separate task.
