+++ backend/main.py
@@ -474,315 +474,12 @@ def create_refresh_token(data: dict, user_id: int = None):
         return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
 
 
-async def get_current_user(
-    token: str = Depends(oauth2_scheme),
-    request: Request = None,
-    db: Session = Depends(get_db)
-) -> User:
-    """Get current user with impersonation support."""
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-
-    actual_user = None
-
-    # Check if token is an API key
-    if token.startswith('sk_') or token.startswith('pk_live_'):
-        token_hash = hashlib.sha256(token.encode()).hexdigest()
-        api_key = db.query(ApiKey).filter(
-            ApiKey.key_hash == token_hash,
-            ApiKey.is_active == True
-        ).first()
-
-        if not api_key:
-            api_key = db.query(ApiKey).filter(
-                ApiKey.key == token,
-                ApiKey.is_active == True
-            ).first()
-            if api_key:
-                api_key = db.query(ApiKey).filter(
-                    ApiKey.id == api_key.id
-                ).with_for_update().first()
-                if api_key and api_key.key is not None:
-                    api_key.key_hash = token_hash
-                    api_key.key_prefix = token[:8]
-                    api_key.key = None
-                    db.commit()
-
-        if api_key is None:
-            raise credentials_exception
-
-        if api_key.expires_at and api_key.expires_at < datetime.now(timezone.utc):
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="API key has expired",
-                headers={"WWW-Authenticate": "Bearer"},
-            )
-
-        api_key.last_used_at = datetime.now(timezone.utc)
-        db.commit()
-
-        actual_user = db.query(User).filter(User.id == api_key.user_id).first()
-        if actual_user is None:
-            raise credentials_exception
-
-        if request:
-            request.state._api_key_obj = api_key
-
-        if request and api_key.scopes:
-            try:
-                from auth.scope_enforcement import check_endpoint_scopes
-                if not check_endpoint_scopes(request, api_key.scopes):
-                    raise HTTPException(
-                        status_code=status.HTTP_403_FORBIDDEN,
-                        detail="API key lacks required scope for this endpoint",
-                    )
-            except ImportError:
-                pass
-            except HTTPException:
-                raise
-
-    else:
-        if _USE_SECURE_TOKENS:
-            token_data = _verify_secure_token(token, expected_type=TokenType.ACCESS)
-            if not token_data:
-                raise credentials_exception
-            email = token_data.sub
-        else:
-            try:
-                _jwt_aud = os.getenv("JWT_AUDIENCE", "perennia-crm")
-                payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM], audience=_jwt_aud, options={"verify_aud": True})
-                _expected_issuer = os.getenv("JWT_ISSUER", "perennia-api")
-                _token_issuer = payload.get("iss")
-                if _token_issuer and _token_issuer != _expected_issuer:
-                    logger.warning(f"JWT issuer mismatch: expected={_expected_issuer}, got={_token_issuer}")
-                    raise credentials_exception
-                email: str = payload.get("sub")
-                if email is None:
-                    raise credentials_exception
-                token_scope = payload.get("scope")
-                if token_scope in ("mfa_verify", "mfa_setup"):
-                    logger.warning(f"Rejected MFA scoped token ({token_scope}) used as access token for {email}")
-                    raise credentials_exception
-                jti = payload.get("jti")
-                if jti:
-                    try:
-                        from auth.tokens import is_token_blacklisted
-                        if is_token_blacklisted(jti):
-                            raise credentials_exception
-                    except ImportError:
-                        pass
-            except InvalidTokenError:
-                raise credentials_exception
-
-        actual_user = db.query(User).filter(User.email == email).first()
-        if actual_user is None:
-            raise credentials_exception
-
-        if not getattr(actual_user, "is_active", True):
-            raise HTTPException(status_code=401, detail="Account deactivated")
-        if getattr(actual_user, "locked_until", None) and actual_user.locked_until > datetime.now(timezone.utc):
-            raise HTTPException(status_code=401, detail="Account temporarily locked")
-
-    # Check for impersonation
-    from auth.dependencies import resolve_impersonation
-    actual_user = resolve_impersonation(request, actual_user, db)
-
-    # Update last_activity_at (throttled to every 5 minutes)
-    try:
-        now = datetime.now(timezone.utc)
-        if actual_user.last_activity_at is None or (now - actual_user.last_activity_at).total_seconds() > 300:
-            actual_user.last_activity_at = now
-            db.commit()
-    except Exception as e:
-        logger.debug(f"Failed to update last_activity_at: {e}")
-        db.rollback()
-
-    return actual_user
-
-
-async def get_current_user_flexible(
-    request: Request,
-    db: Session = Depends(get_db)
-) -> User:
-    """Flexible authentication supporting Bearer token and X-API-Key header."""
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="Not authenticated",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-
-    token = None
-
-    # Check X-API-Key header first
-    api_key_header = request.headers.get("X-API-Key")
-    if api_key_header:
-        header_hash = hashlib.sha256(api_key_header.encode()).hexdigest()
-        api_key = db.query(ApiKey).filter(
-            ApiKey.key_hash == header_hash,
-            ApiKey.is_active == True
-        ).first()
-
-        if not api_key:
-            api_key = db.query(ApiKey).filter(
-                ApiKey.key == api_key_header,
-                ApiKey.is_active == True
-            ).first()
-            if api_key:
-                api_key = db.query(ApiKey).filter(
-                    ApiKey.id == api_key.id
-                ).with_for_update().first()
-                if api_key and api_key.key is not None:
-                    api_key.key_hash = header_hash
-                    api_key.key_prefix = api_key_header[:8]
-                    api_key.key = None
-                    db.commit()
-
-        if api_key:
-            if api_key.expires_at and api_key.expires_at < datetime.now(timezone.utc):
-                raise HTTPException(
-                    status_code=status.HTTP_401_UNAUTHORIZED,
-                    detail="API key has expired",
-                    headers={"WWW-Authenticate": "Bearer"},
-                )
-
-            api_key.last_used_at = datetime.now(timezone.utc)
-            db.commit()
-
-            if hasattr(request, 'state'):
-                request.state._api_key_obj = api_key
-
-            if api_key.scopes:
-                try:
-                    from auth.scope_enforcement import check_endpoint_scopes
-                    if not check_endpoint_scopes(request, api_key.scopes):
-                        raise HTTPException(
-                            status_code=status.HTTP_403_FORBIDDEN,
-                            detail="API key lacks required scope for this endpoint",
-                        )
-                except ImportError:
-                    pass
-                except HTTPException:
-                    raise
-
-            if api_key.organization_id:
-                from database.tenant_mixin import set_tenant_context
-                set_tenant_context(db, api_key.organization_id)
-                logger.info(f"API key tenant context set to org {api_key.organization_id}")
-
-            actual_user = db.query(User).filter(User.id == api_key.user_id).first()
-            if actual_user:
-                from auth.dependencies import resolve_impersonation
-                return resolve_impersonation(request, actual_user, db, auth_method="API key")
-
-        raise credentials_exception
-
-    # Check Authorization header
-    auth_header = request.headers.get("Authorization")
-    if auth_header and auth_header.startswith("Bearer "):
-        token = auth_header.replace("Bearer ", "")
-
-    if not token:
-        raise credentials_exception
-
-    # Check if token is an API key in Bearer header
-    if token.startswith('sk_') or token.startswith('pk_live_'):
-        token_hash = hashlib.sha256(token.encode()).hexdigest()
-        api_key = db.query(ApiKey).filter(
-            ApiKey.key_hash == token_hash,
-            ApiKey.is_active == True
-        ).first()
-
-        if not api_key:
-            api_key = db.query(ApiKey).filter(
-                ApiKey.key == token,
-                ApiKey.is_active == True
-            ).first()
-            if api_key:
-                api_key = db.query(ApiKey).filter(
-                    ApiKey.id == api_key.id
-                ).with_for_update().first()
-                if api_key and api_key.key is not None:
-                    api_key.key_hash = token_hash
-                    api_key.key_prefix = token[:8]
-                    api_key.key = None
-                    db.commit()
-
-        if api_key is None:
-            raise credentials_exception
-
-        if api_key.expires_at and api_key.expires_at < datetime.now(timezone.utc):
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="API key has expired",
-                headers={"WWW-Authenticate": "Bearer"},
-            )
-
-        api_key.last_used_at = datetime.now(timezone.utc)
-        db.commit()
-
-        if hasattr(request, 'state'):
-            request.state._api_key_obj = api_key
-
-        if api_key.scopes:
-            try:
-                from auth.scope_enforcement import check_endpoint_scopes
-                if not check_endpoint_scopes(request, api_key.scopes):
-                    raise HTTPException(
-                        status_code=status.HTTP_403_FORBIDDEN,
-                        detail="API key lacks required scope for this endpoint",
-                    )
-            except ImportError:
-                pass
-            except HTTPException:
-                raise
-
-        actual_user = db.query(User).filter(User.id == api_key.user_id).first()
-        if actual_user is None:
-            raise credentials_exception
-
-        from auth.dependencies import resolve_impersonation
-        return resolve_impersonation(request, actual_user, db, auth_method="Bearer API key")
-
-    # JWT token
-    if _USE_SECURE_TOKENS:
-        token_data = _verify_secure_token(token, expected_type=TokenType.ACCESS)
-        if not token_data:
-            raise credentials_exception
-        email = token_data.sub
-    else:
-        try:
-            _jwt_aud = os.getenv("JWT_AUDIENCE", "perennia-crm")
-            payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM], audience=_jwt_aud, options={"verify_aud": True})
-            _expected_issuer = os.getenv("JWT_ISSUER", "perennia-api")
-            _token_issuer = payload.get("iss")
-            if _token_issuer and _token_issuer != _expected_issuer:
-                logger.warning(f"JWT issuer mismatch: expected={_expected_issuer}, got={_token_issuer}")
-                raise credentials_exception
-            email: str = payload.get("sub")
-            if email is None:
-                raise credentials_exception
-            token_scope = payload.get("scope")
-            if token_scope in ("mfa_verify", "mfa_setup"):
-                logger.warning(f"Rejected MFA scoped token ({token_scope}) used as access token for {email}")
-                raise credentials_exception
-        except InvalidTokenError:
-            raise credentials_exception
-
-    actual_user = db.query(User).filter(User.email == email).first()
-    if actual_user is None:
-        raise credentials_exception
-
-    if not getattr(actual_user, "is_active", True):
-        raise HTTPException(status_code=401, detail="Account deactivated")
-    if getattr(actual_user, "locked_until", None) and actual_user.locked_until > datetime.now(timezone.utc):
-        raise HTTPException(status_code=401, detail="Account temporarily locked")
-
-    # Check for impersonation
-    from auth.dependencies import resolve_impersonation
-    return resolve_impersonation(request, actual_user, db, auth_method="flexible")
+# Back-compat re-exports: real bodies now live in auth/dependencies.py.
+# All existing `from main import get_current_user` calls continue to work.
+from auth.dependencies import (  # noqa: F401
+    get_current_user,
+    get_current_user_flexible,
+)
 
 
 # ============================================================================
