+++ backend/seed_test_account.py
@@ -25,9 +25,11 @@ from decimal import Decimal
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
+import logging
 from database import SessionLocal
-from main import logger
 import bcrypt as _bcrypt
+
+logger = logging.getLogger(__name__)
 from sqlalchemy import text
